+++ contracts/pool/src/contract.rs
@@ -660,6 +660,12 @@ fn do_swap(
     //     }
     // }
 
+    if let Some(max_spread) = max_spread {
+        if !(0..=config.max_allowed_spread_bps).contains(&max_spread) {
+            panic!("max spread is out of bounds")
+        }
+    }
+
     let belief_price = belief_price.map(Decimal::percent);
     let max_spread = Decimal::bps(max_spread.map_or_else(|| config.max_allowed_spread_bps, |x| x));